# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "excerptlab",
    "contact": {}
  },
  "host": "api.excerptlab.com",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admin/listUser": {
      "get": {
        "summary": "listUser",
        "tags": [
          "admin"
        ],
        "operationId": "listUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/user/updateUser/{userId}": {
      "put": {
        "summary": "updateUser",
        "tags": [
          "user"
        ],
        "operationId": "updateUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/submitQuestion": {
      "post": {
        "summary": "submitQuestion",
        "tags": [
          "user"
        ],
        "operationId": "submitQuestion",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/submitQuestionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getProgress": {
      "get": {
        "summary": "getProgress",
        "tags": [
          "user"
        ],
        "operationId": "getProgress",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getUserProfile": {
      "get": {
        "summary": "get user profile",
        "tags": [
          "user"
        ],
        "operationId": "getuserprofile",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getLearningProgress": {
      "get": {
        "summary": "get Learning Progress",
        "tags": [
          "user"
        ],
        "operationId": "getLearningProgress",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "currentDate",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "timePeriod",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/changePassword": {
      "put": {
        "summary": "change password",
        "tags": [
          "user"
        ],
        "operationId": "changepassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/changepasswordrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getDailyStreak": {
      "get": {
        "summary": "getDailyStreak",
        "tags": [
          "user"
        ],
        "operationId": "getDailyStreak",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "currentDate",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/submitFeedback": {
      "post": {
        "summary": "submit Feedback",
        "tags": [
          "user"
        ],
        "operationId": "submitFeedback",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/submitFeedbackRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getLeaderBoardData": {
      "get": {
        "summary": "getLeaderBoardData",
        "tags": [
          "user"
        ],
        "operationId": "getLeaderBoardData",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "currentDate",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "timePeriod",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/user/createNotification": {
      "post": {
        "summary": "create notification",
        "tags": [
          "user"
        ],
        "operationId": "createnotification",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createnotificationrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/listNotification": {
      "get": {
        "summary": "list notification",
        "tags": [
          "user"
        ],
        "operationId": "listnotification",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/deleteNotification/{notiId}": {
      "delete": {
        "summary": "delete notification",
        "tags": [
          "user"
        ],
        "operationId": "deletenotification",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "notiId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/saveFcmToken": {
      "post": {
        "summary": "save fcm token",
        "tags": [
          "user"
        ],
        "operationId": "savefcmtoken",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/savefcmtokenrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "updateUserRequest": {
      "title": "updateUserRequest",
      "example": {
        "name": "xyz",
        "email": "xyz@yopmail.com"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "email"
      ]
    },
    "submitQuestionRequest": {
      "title": "submitQuestionRequest",
      "example": {
        "topicId": "61fa1aabf65f741f0b6c8a95",
        "questionId": "62035c10307021288d59e42e",
        "submittedAnswer": "allegro agetito",
        "submitDate": "2022-02-08T06:36:37.902Z"
      },
      "type": "object",
      "properties": {
        "topicId": {
          "type": "string"
        },
        "questionId": {
          "type": "string"
        },
        "submittedAnswer": {
          "type": "string"
        },
        "submitDate": {
          "type": "string"
        }
      },
      "required": [
        "topicId",
        "questionId",
        "submittedAnswer",
        "submitDate"
      ]
    },
    "changepasswordrequest": {
      "title": "changepasswordrequest",
      "example": {
        "currentPassword": "123456789",
        "newPassword": "123456789"
      },
      "type": "object",
      "properties": {
        "currentPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "required": [
        "currentPassword",
        "newPassword"
      ]
    },
    "submitFeedbackRequest": {
      "title": "submitFeedbackRequest",
      "example": {
        "comment": "music app1",
        "questionId": "622ed17f6d06502821cca315"
      },
      "type": "object",
      "properties": {
        "comment": {
          "type": "string"
        },
        "questionId": {
          "type": "string"
        }
      },
      "required": [
        "comment",
        "questionId"
      ]
    },
    "createnotificationrequest": {
      "title": "createnotificationrequest",
      "example": {
        "receiverId": "61f8f34edcecc51e2466e6d9",
        "message": "notification created"
      },
      "type": "object",
      "properties": {
        "receiverId": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": [
        "receiverId",
        "message"
      ]
    },
    "savefcmtokenrequest": {
      "title": "savefcmtokenrequest",
      "example": {
        "fcmToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MjA5ZjEyNjBiNDZlOTJjM2Y5ZjgzMmQiLCJlbWFpbCI6ImFkbWluQHlvcG1haWwuY29tIiwiaWF0IjoxNjQ0ODI0Mzk4LCJleHAiOjE2NzYzNjAzOTh9.-DrqxPLM2FaMnYyd4b83Xwjc4bRAHrVtxEIJB8G0SrM"
      },
      "type": "object",
      "properties": {
        "fcmToken": {
          "type": "string"
        }
      },
      "required": [
        "fcmToken"
      ]
    }
  },
...
}